+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to NobodysWatching.live are documented here.
 
 ## Added
  - Added new Timezone for Newfoundland
+ - Added Discord Announcements for streamers who are live. (with caveats)
 
 ---